Move from-package text to UML.compartments module
