+++ gaphor/RAAML/fta/notgate.py
@@ -6,14 +6,13 @@ from gaphor.core.modeling import DrawContext
 from gaphor.diagram.presentation import (
     Classified,
     ElementPresentation,
-    text_from_package,
     text_name,
 )
 from gaphor.diagram.shapes import Box, IconBox, stroke
 from gaphor.diagram.support import represents
 from gaphor.RAAML import raaml
 from gaphor.RAAML.fta.constants import DEFAULT_FTA_MAJOR, DEFAULT_FTA_MINOR
-from gaphor.UML.compartments import text_stereotypes
+from gaphor.UML.compartments import text_from_package, text_stereotypes
 
 
 @represents(raaml.NOT)
