+++ gaphor/UML/classes/package.py
@@ -4,12 +4,11 @@ from gaphor import UML
 from gaphor.diagram.presentation import (
     ElementPresentation,
     Named,
-    text_from_package,
     text_name,
 )
 from gaphor.diagram.shapes import Box, cairo_state, stroke
 from gaphor.diagram.support import represents
-from gaphor.UML.compartments import text_stereotypes
+from gaphor.UML.compartments import text_from_package, text_stereotypes
 
 
 @represents(UML.Package)
